begin seperating donations tests into seperate test file
